backend/requirements.txt
pudb
ipython
mysqlclient
django
django-mysql
django-filter
django-cors-headers
djangorestframework
djangorestframework_simplejwt